[XamlC] optimize Datatemplate IL as well
