+++ Xamarin.Forms.Build.Tasks/Xamarin.Forms.Build.Tasks.csproj
@@ -103,6 +103,7 @@
     <Compile Include="CompiledMarkupExtensions\ArrayExtension.cs" />
     <Compile Include="CompiledConverters\ConstraintTypeConverter.cs" />
     <Compile Include="CompiledConverters\ThicknessTypeConverter.cs" />
+    <Compile Include="MethodBodyExtensions.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <Target Name="AfterBuild">
